Reorganize, rename, cleanup files
