+++ src/lib.rs
@@ -6,7 +6,8 @@
 /// COSE algorithms.
 #[allow(non_camel_case_types)]
 pub mod algorithm;
-pub mod claims_set;
+mod claim;
+pub mod cwt;
 mod header_map;
 mod protected;
 /// Implementation of COSE_Sign1.
